refactor : 메소드 이름 변경
- 더 의미 있게 이름 변경
