+++ backend/src/main/java/com/codesquad/coco/room/RoomController.java
@@ -33,6 +33,6 @@ public class RoomController {
 
     @GetMapping("/rooms/{roomId}")
     public RoomDetailDTO findRoomDetail(@PathVariable Long roomId) {
-        return roomService.findRoomDTO(roomId);
+        return roomService.findRoomDetilDTO(roomId);
     }
 }
